Close loop for RFE:2029402 (#11051)
(cherry picked from commit afefffa73593a38cbbc0d1d80d7f10890bdc152d)
